Move scan pixels into cyancia_image
