+++ crates/cyancia_image/build.rs
@@ -8,4 +8,7 @@ fn main() {
         .unwrap()
         .build_artifact()
         .unwrap();
+
+    let compiler = wesl::Wesl::new("src/shaders");
+    compiler.build_artifact(&"package::scan_pixels".parse().unwrap(), "scan_pixels");
 }
