+++ crates/cyancia_image/src/lib.rs
@@ -19,6 +19,7 @@ pub mod blend_modes;
 pub mod composite;
 pub mod dynamic_intermediate_buffer;
 pub mod layer;
+pub mod scan_pixels;
 pub mod texel;
 pub mod tile;
